% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{after}
\alias{after}
\title{Is index after reference ?}
\usage{
after(x, ref)
}
\arguments{
\item{x}{Current index}

\item{ref}{Reference index used to test if x is before or after}
}
\description{
Test if the x index falls before or after the reference index. If it is after, return 1,
             if before, return 0.
}
\examples{
# Is 5 before 2 ?
after(2,5)
# Set result to 0 if the i index is before 5:
a= rep(0,10)
for(i in 1:10){
  a[i]= sqrt(i)*after(i,5)
}

}
\keyword{internal}
