% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Metamodels.R
\name{Allometries}
\alias{Allometries}
\title{Shade tree allometries (optional)}
\usage{
Allometries(S, i)
}
\arguments{
\item{S}{The global list used within the DynACof model.}

\item{i}{The day of interest.}
}
\value{
Any variable that is computed in the function. Default: DM_Stem_Tree, Height_Tree, CrownProj_Tree,
          CrownRad_Tree, Crown_H_Tree, Trunk_H_Tree, LA_Tree and LAD_Tree.
}
\description{
Compute shade tree allometries, which are optional in the model.
             This function is made available to the user to easily change or add equations,
             not to be called directly.
}
\details{
This function is called from the \code{\link{Tree}} parameter functions, and then by the model.
          In-depth details are available in \href{https://goo.gl/NVxcVp}{Vezy (2017)}
}
\references{
Vezy, R., Simulation de pratiques de gestion alternatives pour l’adaptation des plantations pérennes
            aux changements globaux, in École doctorale science de l'environnement, spécialité physique de
            l'environnement. 2017, UNIVERSITÉ DE BORDEAUX: Bordeaux. p. 270.\href{https://goo.gl/NVxcVp}{Link}
}
