% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Disease.R
\name{ALS}
\alias{ALS}
\title{American Leaf Spot}
\usage{
ALS(Elevation, SlopeAzimut = 0, Slope = 0, RowDistance = 1.5, Shade = 0,
  CanopyHeight.Coffee = 2, Fertilization = 3, ShadeType = 5,
  CoffeePruning = c("tree", "row", "block", NULL), df_rain)
}
\arguments{
\item{Elevation}{Site elevation       (m.a.s.l)}

\item{SlopeAzimut}{Slope azimuth        (degree)}

\item{Slope}{Slope percentage     (\eqn{\%})}

\item{RowDistance}{Coffee rows distance (m)}

\item{Shade}{Shade percentage     (\eqn{\%})}

\item{CanopyHeight.Coffee}{Coffee Height        (m)}

\item{Fertilization}{N fertilization per year}

\item{ShadeType}{Shade type:\describe{
\item{1}{Legume only}
\item{2}{Bananas and legume}
\item{3}{Bananas and other plants}
\item{4}{Fruit and forest trees only}
\item{5}{No shade (Full sun)}
}}

\item{CoffeePruning}{Character specifying the pruning management.
Values: \code{tree}, \code{row}, \code{block} or \code{NULL}.}

\item{df_rain}{Data frame with DOY, year and Rain (mm) values}
}
\value{
\item{ALS}{Percentage of dead leaves by ALS by day (\eqn{\%  day^{-1}})}
}
\description{
Compute the percentage of \emph{Coffea} leaves dying from
             American Leaf Spot disease, drought excluded.
}
\note{
It is good practice to use shade tree transmittance to compute \code{Shade} percentage
         (\eqn{Shade= 1-Transmittance}).
}
\examples{
# df_rain has this structure :
df_rain= data.frame(DOY= 1:365, year= rep(2018,365), Rain= rnorm(n = 365,mean = 5, sd = 1))
ALS(Elevation = 1000, df_rain= df_rain)

}
\references{
Avelino et al. (2007) Topography and Crop Management Are Key Factors
for the Development of American Leaf Spot Epidemics on Coffee in Costa Rica
File: "Ia_digitized from Avelino 2007 - JA_lineaire.xlsx"
}
