% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aquiares}
\alias{Aquiares}
\title{Aquiares meteorology.}
\format{A data frame with sixteen variables: \code{year}, \code{DOY},
  \code{Date}, \code{Rain}, \code{Tair}, \code{RH}, \code{RAD}, \code{Pressure},
  \code{WindSpeed}, \code{CO2}, \code{DegreeDays}, \code{Rain}, \code{PAR},
  \code{FDiff}, \code{VPD}, \code{Tmax} and \code{Tmin}. All units corresponds to
  the \code{\link{Meteorology}} function recommendantions.}
\usage{
Aquiares
}
\description{
Aquiares meteorology data at daily time-step
}
\keyword{datasets}
