% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-balance_model.R
\name{balance_model}
\alias{balance_model}
\title{Energy balance}
\usage{
balance_model(S, i)
}
\arguments{
\item{S}{The simulation list}

\item{i}{The day index.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Computes the different components of the energy balance considering the shade tree,
the coffee and the soil.
}
