% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-bud_init.R
\name{bud_init_period}
\alias{bud_init_period}
\title{Bud induction window computation}
\usage{
bud_init_period(S)
}
\arguments{
\item{S}{The list of simulation}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Bud induction can start only at F_Tffb degree-days after vegetative growth stops (Rodriguez et al., 2011).
The following function finds the vegetative growth end day, and add the F_Tffb parameter
(Time of first floral buds, in dd), then find the very first flowering of the year
and set the vector BudInitPeriod to TRUE between the two dates. So buds will appear
between plant F_Tffb parameter and the first flowering day only.
}
\references{
Rodríguez, D., Cure, J., Cotes, J., Gutierrez, A. and Cantor, F., 2011. A coffee agroecosystem model: I. Growth and development of
the coffee plant. Ecological Modelling, 222(19): 3626-3639.
}
