% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{CN}
\alias{CN}
\title{Temperature-dependent correction coefficient for nodes (CB)}
\usage{
CN(x)
}
\arguments{
\item{x}{The average air temperature during the vegetative growing period}
}
\value{
The correction coefficient to compute the number of green nodes in the coffee (see Eq. 27 from Vezy et al. (in prep.))
}
\description{
Compute the temperature-dependent correction coefficient for green nodes in the
coffee plant according to Drinnan and Menzel (1995).
}
\examples{
CN(25)
}
\references{
\itemize{
  \item Drinnan, J. and C. Menzel, Temperature affects vegetative growth and flowering of coffee (Coffea arabica L.).
  Journal of Horticultural Science, 1995. 70(1): p. 25-34.
}
}
