% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-coffee_model.R
\name{coffee_model}
\alias{coffee_model}
\title{Coffee model}
\usage{
coffee_model(S, i)
}
\arguments{
\item{S}{The simulation list}

\item{i}{The day index.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Computes the different components of the coffee growth and yield.
}
