% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Constants.R
\name{Constants}
\alias{Constants}
\title{Constants used in the DynACof package}
\usage{
Constants()
}
\value{
This function defines the following constants:
\item{Cp}{specific heat of air for constant pressure (MJ K-1 kg-1), Source: Allen 1998 FAO Eq. 8 p. 32}
\item{pressure0}{reference atmospheric pressure at sea level (Pa)}
\item{FPAR}{Fraction of global radiation that is PAR (source: MAESPA model)}
\item{g}{gravitational acceleration (m s-2)}
\item{Rd}{gas constant of dry air (J kg-1 K-1), source : Foken p. 245}
\item{Rgas}{universal gas constant (J mol-1 K-1)}
\item{Kelvin}{conversion degree Celsius to Kelvin}
\item{vonkarman}{von Karman constant (-)}
\item{MJ_to_W}{coefficient to convert MJ into W (W MJ-1)}
\item{Gsc}{solar constant (W m-2, = J m-2 s-1), source : Khorasanizadeh and Mohammadi (2016)}
\item{sigma}{Stefan-Boltzmann constant (W m-2 K-4)}
\item{H2OMW}{Conversion factor from kg to mol for H2O (kg mol-1)}
\item{W_umol}{Conversion factor from watt to micromole for H2O (W umol-1)}
\item{\eqn{\lambda} (lambda)}{Latent heat of vaporization (MJ kgH2O-1)}
}
\description{
Constants used in the DynACof package
}
\note{
Values are partly burrowed from \code{\link[bigleaf]{bigleaf.constants}}
}
\seealso{
\code{\link{DynACof}}
}
