% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{cos_deg}
\alias{cos_deg}
\alias{sin_deg}
\alias{tan_deg}
\alias{acos_deg}
\alias{asin_deg}
\alias{atan_deg}
\alias{sin_deg}
\alias{tan_deg}
\alias{acos_deg}
\alias{asin_deg}
\alias{atan_deg}
\title{Trigonometric Functions (degree angles version)}
\usage{
cos_deg(x)

sin_deg(x)

tan_deg(x)

acos_deg(x)

asin_deg(x)

atan_deg(x)
}
\arguments{
\item{x}{Angle in degree}
}
\description{
These functions give the obvious trigonometric functions.
             They respectively compute the cosine, sine, tangent,
             arc-cosine, arc-sine, arc-tangent as in the base functions
             \code{\link[base]{Trig}}, but use input in degree instead
             of radian.
}
\details{
The conversions between radian to degree is:

\deqn{x*pi/180}
}
\examples{
# cosinus of an angle of 120 degree:
cos_deg(120)
# should yield -0.5, as in the base version:
cos(120*pi/180)

}
