% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-dynacof_julia.R
\name{dynacof.jl}
\alias{dynacof.jl}
\title{Run DynACof.jl (Julia version of DynACof)}
\usage{
dynacof.jl(Period = NULL, WriteIt = FALSE, Inpath = NULL,
  output_f = ".RData", Outpath = Inpath, Simulation_Name = "DynACof",
  FileName = list(Site = "site.jl", Meteo = "meteorology.txt", Soil =
  "soil.jl", Coffee = "coffee.jl", Tree = NULL), ...)
}
\arguments{
\item{Period}{Period of time to be simulated, see details. Default: \code{NULL}}

\item{WriteIt}{If \code{TRUE}, write the outputs to disk using \code{\link[=write.results]{write.results()}}, see details. Default: \code{FALSE}}

\item{Inpath}{Path to the input parameter list folder, Default: \code{NULL} (take package values)}

\item{output_f}{Output format. If \code{output_f = ".RData"}, the output list will be saved as a unique \code{.RData} file. Any other value:
write the output list in several \code{.csv} and \code{.txt} files. Default: \code{.RData}}

\item{Outpath}{Path pointing to the folder were the results will be writen, Default: \code{Outpath = Inpath}}

\item{Simulation_Name}{Character name of the simulation file name if \code{WriteIt = T}. Default: \code{"DynACof"}}

\item{FileName}{A list of input file names :
\describe{
\item{Site}{Site parameters file name, see details. Default: \code{'site.jl'}}
\item{Meteo}{Meteo parameters file name, see details. Default: \code{'meteorology.txt'}}
\item{Soil}{Soil parameters file name, see details. Default: \code{'soil.jl'}}
\item{Coffee}{Coffee parameters file name, see details. Default: \code{'coffee.jl'}}
\item{Tree}{Shade tree parameters file name, see details. Default: \code{NULL}}
}}

\item{...}{Further arguments to pass to \code{\link[=write.results]{write.results()}}.}
}
\value{
Return invisibly a list containing three objects (Parameters, Meteo and Sim). See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Make a simulation using the julia version of DynACof, \href{https://github.com/VEZY/DynACof.jl}{DynACof.jl}.
This implementation is much faster, with run-times ~100x faster. Need Julia installed on the computer and in the path.
}
\note{
If the user needs default values from the package for FileName names, put their values to "package". Careful:
Always call \code{\link[=dynacof.jl_setup]{dynacof.jl_setup()}} before runnning \code{dynacof.jl}. See example.
}
\examples{
\dontrun{
# Setting up julia + DynACof.jl:
dynacof.jl_setup()

# Running a simulation:
# First, downloading an example meteorology file:
met_file= tempfile()
download.file("https://raw.githubusercontent.com/VEZY/DynACof.jl_inputs/master/meteorology.txt",met_file)


S= dynacof.jl(Inpath= dirname(met_file), FileName=list(Site="package",Meteo=basename(met_file),Soil="package",Coffee="package",Tree=NULL))
unlink(met_file)
# Plotting the coffee LAI along the simulation:
plot(S$Meteo$Date,S$Sim$LAI)
}
}
