% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-dynacof_julia.R
\name{dynacof.jl_update}
\alias{dynacof.jl_update}
\title{Update DynACof.jl}
\usage{
dynacof.jl_update()
}
\description{
Update the \href{https://github.com/VEZY/DynACof.jl}{DynACof.jl} julia package.
}
\examples{
\dontrun{
# Can be time-consuming.
DynACof::dynacof.jl_update()
}
}
