% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Coffee_DCM.R
\name{DynACof}
\alias{DynACof}
\title{Dynamic Agroforestry Coffee Crop Model}
\usage{
DynACof(Period = NULL, WriteIt = F, returnIt = F, ...,
  output_f = ".RData", Inpath = NULL, Outpath = Inpath,
  Site = "1-Site.R", Meteo = "2-Meteorology.txt", Soil = "3-Soil.R",
  Coffee = "4-Coffee.R", Tree = NULL)
}
\arguments{
\item{Period}{Period of time to be simulated, see details. Default: \code{NULL}}

\item{WriteIt}{If \code{TRUE}, write the resulting list, see details. Default: \code{FALSE}}

\item{returnIt}{Are the results to be returned by the function ? Default: \code{FALSE}}

\item{...}{PARAM_DESCRIPTION}

\item{output_f}{Output format, if '.RData', the output list will be saved as a unique \code{.RData} file, if it is different from
\code{.RData}, the output list will be writen in several \code{.csv} and \code{.txt} format. Default: \code{.RData}}

\item{Inpath}{Path to the input parameter list folder, Default: \code{"1-Input/Default"}}

\item{Outpath}{Path pointing to the folder were the results will be writen, Default: \code{=Inpath}}

\item{Site}{Site parameters file name, see details. Default: \code{'1-Site.R'}}

\item{Meteo}{Meteo parameters file name, see details. Default: \code{'2-Meteorology.txt'}}

\item{Soil}{Soil parameters file name, see details. Default: \code{'3-Soil.R'}}

\item{Coffee}{Coffee parameters file name, see details. Default: \code{'4-Coffee.R'}}

\item{Tree}{Shade tree parameters file name, see details. Default: \code{NULL}}
}
\value{
A list containing three objects :
\itemize{
  \item A data.frame of the simulation outputs at daily time-step:  #' \tabular{llll}{
\strong{Type} \tab \strong{Var} \tab \strong{unit} \tab \strong{Definition}\cr
General                      \tab Cycle                    \tab -                   \tab Plantation cycle ID                                                                \cr
                             \tab Plot_Age                 \tab year                \tab Plantation age                                                                     \cr
                             \tab Plot_Age_num             \tab year (numeric)      \tab Numeric age of plantation                                                          \cr
                             \tab LAIplot                  \tab m2 leaves m-2 soil  \tab Plot (Coffee + Shade Tree if any) Leaf Area Index                                  \cr
Organs Coffee                \tab *_RE                     \tab -                   \tab Reserves                                                                           \cr
                             \tab *_SCR                    \tab -                   \tab Stump and Coarse roots                                                             \cr
                             \tab *_Fruit                  \tab -                   \tab Fruit                                                                              \cr
                             \tab *_RsWood                 \tab -                   \tab Resprout wood (= branches)                                                         \cr
                             \tab *_FRoot                  \tab -                   \tab Fine roots                                                                         \cr
                             \tab *_Leaf                   \tab                     \tab Leaves                                                                             \cr
Organs shade tree            \tab *_RE_Tree                \tab -                   \tab Reserves                                                                           \cr
                             \tab *_Stem_Tree              \tab -                   \tab Stem (= trunk)                                                                     \cr
                             \tab *_Branch_Tree            \tab -                   \tab Branches                                                                           \cr
                             \tab *_CoarseRoot_Tree        \tab -                   \tab Coarse roots                                                                       \cr
                             \tab *_FRoot_Tree             \tab -                   \tab Fine roots                                                                         \cr
                             \tab *_Leaf_Tree              \tab                     \tab Leaves                                                                             \cr
Energy                       \tab Rn_tot                   \tab MJ m-2 d-1          \tab System net radiation                                                               \cr
                             \tab Rn_Tree                  \tab MJ m-2 d-1          \tab Shade tree net radiation                                                           \cr
                             \tab Rn_Coffee                \tab MJ m-2 d-1          \tab Coffee net radiation                                                               \cr
                             \tab Rn_Soil                  \tab MJ m-2 d-1          \tab Soil net radiation                                                                 \cr
                             \tab LE_*                     \tab MJ m-2 d-1          \tab System/Coffee/Tree/Soil latent heat                                                \cr
                             \tab H_*                      \tab MJ m-2 d-1          \tab System/Coffee/Tree/Soil sensible heat                                              \cr
                             \tab Q_Soil                   \tab MJ m-2 d-1          \tab Soil heat transport                                                                \cr
                             \tab Transmittance_Tree       \tab fraction            \tab Light transmitted under the shade trees canopy                                     \cr
                             \tab K_Dir                    \tab -                   \tab Direct light extinction coefficient                                                \cr
                             \tab K_Dif                    \tab -                   \tab Diffuse light extinction coefficient                                               \cr
                             \tab APAR                     \tab MJ m-2 d-1          \tab Absorbed PAR by the plant                                                          \cr
                             \tab APAR_Dif                 \tab MJ m-2 d-1          \tab Absorbed diffuse PAR (Direct is APAR-APAR_Dif)                                     \cr
                             \tab lue                      \tab gC MJ               \tab Light use efficiency                                                               \cr
                             \tab Tcan_MAESPA_Coffee       \tab deg C               \tab Coffee canopy temperature computed using MAESPA metamodel                          \cr
                             \tab Tcan_Coffee              \tab deg C               \tab Coffee canopy temperature computed by DynACof                                      \cr
                             \tab DegreeDays_Tcan          \tab deg C               \tab Growing degree days computed using Coffee Canopy Temperature                       \cr
Carbon                       \tab GPP                      \tab gC m-2 d-1          \tab Gross primary productivity                                                         \cr
                             \tab Consumption_RE           \tab gC m-2 d-1          \tab Daily reserve consumption                                                          \cr
                             \tab Carbon_Lack_Mortality    \tab gC m-2 d-1          \tab Mortality from a higher carbon consumption than carbon offer                       \cr
                             \tab Rm                       \tab gC m-2 d-1          \tab Total Coffee maintenance respiration                                               \cr
                             \tab Rm_*                     \tab gC m-2 d-1          \tab Maintenance respiration at organ scale                                             \cr
                             \tab Rc                       \tab gC m-2 d-1          \tab Total Coffee growth respiration                                                    \cr
                             \tab Rc_*                     \tab gC m-2 d-1          \tab Growth respiration at organ scale                                                  \cr
                             \tab Ra                       \tab gC m-2 d-1          \tab Coffee layer autotrophic respiration (=maintenance+growth)                         \cr
                             \tab Demand_*                 \tab gC m-2 d-1          \tab C demand at organ scale (fruit, leaf and fine root only)                           \cr
                             \tab Alloc_*                  \tab gC m-2 d-1          \tab C allocation to organ net of Rm (NPP+Rc)                                           \cr
                             \tab lambdaSCRage             \tab gC gC-1             \tab Age related allocation coefficient to Stump and Coars Roots                        \cr
                             \tab Offer                    \tab gC m-2 d-1          \tab C offer at the begining of the day at layer scale (GPP+Reserve consumption-Rm)     \cr
                             \tab Offer_*                  \tab gC m-2 d-1          \tab C offer to organ net of Rm                                                         \cr
                             \tab Demand_*                 \tab gC m-2 d-1          \tab Total C demand from organ                                                          \cr
                             \tab NPP                      \tab gC m-2 d-1          \tab Net primary productivity at layer scale                                            \cr
                             \tab NPP_*                    \tab gC m-2 d-1          \tab Net primary productivity at organ scale                                            \cr
                             \tab Mnat_*                   \tab gC m-2 d-1          \tab Organ natural mortality (= due to lifespan)                                        \cr
                             \tab Mprun_*                  \tab gC m-2 d-1          \tab Organ mortality due to pruning                                                     \cr
                             \tab MALS                     \tab gC m-2 d-1          \tab Coffee leaf mortality from American Leaf Spot                                      \cr
                             \tab Mortality_*              \tab gC m-2 d-1          \tab Total organ mortality                                                              \cr
                             \tab LAI                      \tab m2 leaves m-2 soil  \tab Leaf Area Index                                                                    \cr
                             \tab CM_*                     \tab gC m-2 d-1          \tab Organ C mass                                                                       \cr
                             \tab DM_*                     \tab gDM m-2 d-1         \tab Organ dry mass                                                                     \cr
                             \tab Height_Tree              \tab m                   \tab Shade tree total height (used for boundary conductance), set to 0 if no shade trees\cr
Fruit development            \tab BudInitPeriod            \tab boolean             \tab Bud initiation period                                                              \cr
                             \tab Budinit                  \tab Buds d-1            \tab Total Number of Buds Initiated per day                                             \cr
                             \tab ratioNodestoLAI          \tab Nodes LAI-1         \tab Number of fruiting nodes per LAI unit                                              \cr
                             \tab Temp_cor_Bud             \tab fraction            \tab Temperature correction factor for bud development                                  \cr
                             \tab p_budbreakperday         \tab 0-1                 \tab Daily probability of bud dormancy break                                            \cr
                             \tab BudBreak                 \tab Buds d-1            \tab Total number of buds breaking dormancy per day                                     \cr
                             \tab Sucrose_Mass             \tab g m-2 d-1           \tab Coffee Fruit Sucrose Mass                                                          \cr
                             \tab Sucrose_Content          \tab g Sugar gDM         \tab Coffee Fruit Sucrose Content                                                       \cr
                             \tab Maturation_duration    \tab days Fruit cohort-1 \tab Coffee Fruit Total Maturation Duration for each cohort                             \cr
                             \tab Harvest_Maturity_Pot     \tab Fraction            \tab Daily average fruit maturity (0-1)                                                 \cr
                             \tab Date_harvest             \tab day of year         \tab date of harvest                                                                    \cr
                             \tab Harvest_Fruit            \tab gC m-2              \tab Total fruit carbon mass at harvest                                                 \cr
                             \tab Harvest_Maturity         \tab Fraction            \tab Average fruit maturity at harvest (0-1)                                            \cr
                             \tab Overriped_Fruit          \tab gC m-2 d-1          \tab Overriped fruits that fall onto the ground                                         \cr
Water                        \tab IntercMax                \tab mm                  \tab Maximum potential rainfall interception by canopy                                  \cr
                             \tab CanopyHumect             \tab mm                  \tab Rainfall interception by canopy                                                    \cr
                             \tab Throughfall              \tab mm                  \tab Rainfall not intercepted by the canopy, coming to the soil                         \cr
                             \tab SuperficialRunoff        \tab mm                  \tab Water runoff from the superficial soil layer                                       \cr
                             \tab ExcessRunoff             \tab mm                  \tab Discharge from the superficial soil layer                                          \cr
                             \tab TotSuperficialRunoff     \tab mm                  \tab Sum of discharge+ExcessRunoff                                                      \cr
                             \tab InfilCapa_mmd            \tab mm                  \tab Superficial water infiltration capacity to first layer of soil                     \cr
                             \tab Infiltration             \tab mm                  \tab Superficial water infiltration to first layer of soil                              \cr
                             \tab Drain_[1-3]              \tab mm                  \tab Water drainage from soil layer 1, 2 or 3                                           \cr
                             \tab WSurfaceRes              \tab mm                  \tab Soil water content from the surface layer                                          \cr
                             \tab W_tot                    \tab mm                  \tab Total soil profile water content                                                   \cr
                             \tab W_[1-3]                  \tab mm                  \tab Soil water content from the layer 1, 2 or 3                                        \cr
                             \tab REW_tot                  \tab -                   \tab Relative extractable water from the soil                                           \cr
                             \tab REW_[1-3]                \tab -                   \tab Relative extractable water from the layer 1, 2 or 3                                \cr
                             \tab EW_tot                   \tab mm                  \tab Extractable water from the soil                                                    \cr
                             \tab EW_[1-3]                 \tab mm                  \tab Extractable water from the layer 1, 2 or 3                                         \cr
                             \tab SWD                      \tab mm                  \tab soil water deficit                                                                 \cr
                             \tab RootWaterExtract_[1-3]   \tab mm                  \tab Root water extraction for soil layer 1 to 3                                        \cr
                             \tab IntercRevapor            \tab mm                  \tab Evaporation by canopy                                                              \cr
                             \tab T_*                      \tab mm                  \tab Transpiration at system/Coffee/Tree scale                                          \cr
                             \tab E_Soil                   \tab mm                  \tab Soil evaporation                                                                   \cr
                             \tab ETR                      \tab mm                  \tab System evapotranspiration                                                          \cr
                             \tab SoilWaterPot             \tab MPa                 \tab Soil water potential                                                               \cr
                             \tab LeafWaterPotential       \tab Mpa                 \tab Coffee leaf water potential                                                        \cr
Special shade tree variables \tab LA_Tree                  \tab m2 leaves tree-1    \tab shade tree leaf area                                                               \cr
                             \tab Crown_H_Tree             \tab m                   \tab Crown height                                                                       \cr
                             \tab Trunk_H_Tree             \tab m                   \tab Trunk height                                                                       \cr
                             \tab DBH_Tree_m               \tab m                   \tab Diameter at breast height                                                          \cr
                             \tab LAD_Tree                 \tab m2 m-3              \tab Shade tree Leaf Area Density                                                       \cr
                             \tab CrownRad_Tree            \tab m                   \tab Crown radius                                                                       \cr
                             \tab CrownProj_Tree           \tab m2 crown tree-1     \tab Crown projection                                                                   \cr
                             \tab Stocking_Tree            \tab tree m-2            \tab Shade tree density                                                                 \cr
                             \tab TimetoThin_Tree          \tab boolean             \tab Is it time to thin the shade tree                                                  \cr
                             \tab MThinning_*_Tree         \tab gc m-2 d-1          \tab Mortality due to thining at organ scale
}

  \item A data.frame of the input meteorology, potentially coming from the output of \code{\link{Meteorology}}: \tabular{lll}{\strong{Var} \tab \strong{unit} \tab \strong{Definition}\cr
                    year            \tab year        \tab Year of the simulation                       \cr
                    DOY             \tab day         \tab day of the year                              \cr
                    Date            \tab POSIXct date\tab Date in POSICct format                       \cr
                    Rain            \tab mm          \tab Rainfall                                     \cr
                    Tair            \tab deg C       \tab Air temperature (above canopy)               \cr
                    RH              \tab \%          \tab Relative humidity                            \cr
                    RAD             \tab MJ m-2 d-1  \tab Incident shortwave radiation                 \cr
                    Pressure        \tab hPa         \tab Atmospheic pressure                          \cr
                    WindSpeed       \tab m s-1       \tab Wind speed                                   \cr
                    CO2             \tab ppm         \tab Atmospheric CO2 concentration                \cr
                    DegreeDays      \tab deg C       \tab Growing degrre days                          \cr
                    PAR             \tab MJ m-2 d-1  \tab Incident photosynthetically active radiation \cr
                    FDiff           \tab Fraction    \tab Diffuse light fraction                       \cr
                    VPD             \tab hPa         \tab Vapor pressure deficit                       \cr
                    Rn              \tab MJ m-2 d-1  \tab Net radiation (will be removed further)      \cr
                    Tmax            \tab deg C       \tab Maximum air temperature durnig the day       \cr
                    Tmin            \tab deg C       \tab Minimum air temperature durnig the day       \cr
                    rho             \tab kg m-3      \tab Air density of moist air                     \cr
                    DaysWithoutRain \tab day         \tab Number of consecutive days with no rainfall
}
  \item A list of the input parameters (see \code{\link{site}})
}
}
\description{
The DynACof process-based model computes plot-scale Net Primary Productivity, carbon allocation, growth, yield,
               energy, and water balance of coffee plantations according to management, while accounting for spatial effects using
               metamodels from the 3D process-based MAESPA. The model also uses coffee bud and fruit cohorts for reproductive
               development to better represent fruit carbon demand distribution along the year.
}
\details{
Almost all variables for coffee exist also for shade trees with a suffix
         \code{_Tree} after the name of the variable, e.g. : LAI = coffee LAI,
         LAI_Tree = shade tree LAI.
         Special shade tree variables (see return section) are only optional,
         and may be longer upon parameterization because variables can be added in
         the metamodels parameter file in \strong{\code{\link{Metamodels}}} or
         \strong{\code{\link{Allometries}}}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[bigleaf]{aerodynamic.conductance}}
\code{\link{Meteorology}} \code{\link{site}}
}
\author{
R. Vezy; O. Roupsard
}
