% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-dynacof_julia.R
\name{dynacof_i.jl}
\alias{dynacof_i.jl}
\title{Step-by-step \code{\link[=dynacof.jl]{dynacof.jl()}}}
\usage{
dynacof_i.jl(i, S)
}
\arguments{
\item{i}{Either an integer, or a range giving the day of simulation needed. Match the row index, so \code{i=1} make
a simulation for the first row of Sim and Met.}

\item{S}{The simulation list, output of \code{\link[=dynacof.jl]{dynacof.jl()}}. It is mandatory that \code{S} come from a \code{\link[=dynacof.jl]{dynacof.jl()}} simulation because
the parameters linked to a simulation have either \code{R} or \code{Julia} code affiliated to the simulation.}
}
\value{
The modified simulation list \code{S}
}
\description{
Using DynACof one iteration after another using the julia version of DynACof, \href{https://github.com/VEZY/DynACof.jl}{DynACof.jl}.
Allows to run a DynACof simulation with starting at age > 0 with initializations.
}
\section{Fields}{

\describe{
\item{\code{Careful}}{Always call \code{\link[=dynacof.jl_setup]{dynacof.jl_setup()}} before runnning \code{dynacof_i.jl}. See example.}
}}

\examples{
\dontrun{
# Setting up julia + DynACof.jl:
dynacof.jl_setup()

Making a regular simulation using example data:
# First, downloading an example meteorology file:
met_file= tempfile()
m= "https://raw.githubusercontent.com/VEZY/DynACof.jl_inputs/master/meteorology.txt"
download.file(m,met_file)


S= dynacof.jl(Inpath= dirname(met_file), FileName=list(Site="package",Meteo=basename(met_file),
                                                       Soil="package",Coffee="package",Tree=NULL))
unlink(met_file)

# Value of the maintenance respiration for coffee on day i=100:
i= 100
S$Sim$Rm[i]

# Changing the value of Tair in the meteorology for day 100:
S$Meteo$Tair[i]= S$Meteo$Tair[i]+10.0

S= dynacof_i(i,S)

# New value of the maintenance respiration for coffee:
S$Sim$Rm[i]

# To re-run DynACof for several days, use a range for `i`:
S= dynacof_i(i:(i+10),S)
}
}
