% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-energy_model.R
\name{energy_model_coffee}
\alias{energy_model_coffee}
\alias{energy_model_tree}
\title{Energy fluxes models}
\usage{
energy_model_coffee(S, i)

energy_model_tree(S, i)
}
\arguments{
\item{S}{The simulation list}

\item{i}{The day index.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Computes the energy-related variables such as H, LE, Tleaf for the shade tree or the coffee.
}
\seealso{
\code{\link[=energy_water_models]{energy_water_models()}}
}
