% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-energy_model.R
\name{energy_water_models}
\alias{energy_water_models}
\title{Energy and water models}
\usage{
energy_water_models(S, i)
}
\arguments{
\item{S}{The simulation list}

\item{i}{The day index.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Computes the energy and water related variables for the shade tree (if any), the coffee
and the soil. Call different sub-models:
\itemize{
\item \code{\link[=light_model_tree]{light_model_tree()}} for the light interception of the shade tree
\item \code{\link[=light_model_coffee]{light_model_coffee()}} for the light interception of the coffee
\item \code{\link[=energy_model_tree]{energy_model_tree()}} for the energy fluxes of the tree (H, LE, Tleaf...)
\item \code{\link[=energy_model_coffee]{energy_model_coffee()}} for the energy fluxes of the coffee (H, LE, Tleaf...)
\item \code{\link[=soil_model]{soil_model()}} the full soil model (water transport, H, T Soil...)
\item \code{\link[=balance_model]{balance_model()}} the energy balance at plot scale model (H, LE, Rn...)
}
}
