% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{F_repartition}
\alias{F_repartition}
\alias{F_densite}
\alias{F_Integ_Dens}
\title{Logistic function helpers}
\usage{
F_repartition(xi, u_log, s_log)

F_densite(xi, u_log, s_log)

F_Integ_Dens(x, index, u_log, s_log)
}
\arguments{
\item{xi}{X vector}

\item{u_log}{Inflexion point (x-value of the sigmoid's midpoint)}

\item{s_log}{Steepness of the curve}

\item{index}{Index at which to compute the delta between x's}
}
\value{
\item{F_repartition}{Logistic function distibution}
        \item{F_densite}{Logistic function density}
        \item{F_Integ_Dens}{Delta between logistic function x's}
}
\description{
Compute the distribution, density and delta of the
             logistic function
}
\examples{
F_repartition(1:10,5,0.1)
F_densite(1:10,5,0.1)
F_Integ_Dens(1:10,10,5,0.1)

}
\seealso{
More informations can be found in \href{https://en.wikipedia.org/wiki/Logistic_function}{
            the wikipedia page}
}
\keyword{internal}
