% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{G_interlay}
\alias{G_interlay}
\title{Aerodynamic conductance between layers of canopy}
\usage{
G_interlay(Wind, ZHT, Z1, Z2, Z0 = Z2 * 0.1, ZPD = Z2 * 0.75, alpha = 1.5,
  ZW = ZPD + alpha * (Z2 - ZPD), LAI, extwind = 0,
  vonkarman = Constants()$vonkarman)
}
\arguments{
\item{Wind}{Average daily wind speed above canopy (m s-1)}

\item{ZHT}{Wind measurement height (m)}

\item{Z1}{Average canopy height of the shorter crop (m)}

\item{Z2}{Average canopy height of the taller crop (m)}

\item{Z0}{Roughness length (m). Default: \code{0.1*TREEH}}

\item{ZPD}{Zero-plane displacement (m), Default: \code{0.75*TREEH}}

\item{alpha}{Constant for diffusivity at top canopy. Default: \code{1.5} following
Van de Griend et al (1989).}

\item{ZW}{Top height of the roughness sublayer (m). Default: \code{ZPD+alpha*(Z2-ZPD)}}

\item{LAI}{Leaf area index of the upper layer (m2 leaf m-2 soil).}

\item{vonkarman}{Von Karman constant, default to \code{Constants()$vonkarman}, 0.41.}
}
\value{
\item{g_af}{The aerodynamic conductance of the air between two canopy layers (m s-1)}
}
\description{
Compute the aerodynamic conductance between canopy layers
             following Van de Griend and Van Boxel (1989).
}
\details{
\code{alpha} can also be computed as:
         \deqn{alpha=\frac{zw-d}{Z2-d}}{alpha= (zw-d)/(Z2-d)}
         The aerodynamic conductance between canopy layers is computed as:
         \deqn{g_{af}= \frac{1}{\frac{U_h}{K_h}\ln(U_{mid}/U_{inter})}}{g_af= 1/((Uh/Kh)*log(U_mid/U_inter))}
         where \eqn{U_mid} is the wind speed at median cumulated LAI between the top and the soil, and
         \eqn{U_inter} is the wind speed at height between the two canopy layers.
}
\examples{
# G_af for a coffee plantation managed in agroforestry system:
G_interlay(Wind=3,ZHT=25,Z1= site()$Height_Coffee,Z2=24,LAI= 0.5,extwind=0.58)


}
\references{
Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model
            with a multilayer canopy representation for remote sensing purposes. Water
            Resources Research, 1989. 25(5): p. 949-971.
}
\seealso{
\code{\link{G_bulk}} and \code{\link{GetWind}}, which is used internaly.
}
