% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{G_soilcan}
\alias{G_soilcan}
\title{Canopy to soil aerodynamic conductance}
\usage{
G_soilcan(Wind, ZHT, Z_top, Z0 = Z_top * 0.1, ZPD = Z_top * 0.75,
  alpha = 1.5, ZW = ZPD + alpha * (Z_top - ZPD), LAI_top, LAI_bot,
  extwind = 0, vonkarman = Constants()$vonkarman)
}
\arguments{
\item{Wind}{Average daily wind speed above canopy (m s-1)}

\item{ZHT}{Wind measurement height (m)}

\item{Z_top}{Average canopy height of the taller crop (m)}

\item{Z0}{Roughness length (m). Default: \code{0.1*Z_top}}

\item{ZPD}{Zero-plane displacement (m), Default: \code{0.75*Z_top}}

\item{alpha}{Constant for diffusivity at top canopy. Default: \code{1.5} following
Van de Griend et al (1989).}

\item{ZW}{Top height of the roughness sublayer (m). Default: \code{ZPD+alpha*(Z_top-ZPD)}}

\item{LAI_top}{Leaf area index of the upper layer (m2 leaf m-2 soil).}

\item{LAI_bot}{Leaf area index of the layer below the upper layer (m2 leaf m-2 soil).}

\item{extwind}{Extinction coefficient. Default: \code{0}, no extinction.}

\item{vonkarman}{Von Karman constant, default to \code{Constants()$vonkarman}, 0.41.}
}
\value{
\item{\eqn{g_a0}}{The aerodynamic conductance of the air between the lowest canopy layer
                    and the soil surface (m s-1)}
}
\description{
Compute the aerodynamic conductance for sensible and
             latent heat between the center of the lowest canopy layer
             and the soil surface following Van de Griend and Van Boxel (1989).
}
\details{
\code{alpha} can also be computed as:
         \deqn{alpha=\frac{zw-d}{Z_{top}-d}}{alpha= (zw-d)/(Z_top-d)}
         The aerodynamic conductance between the lowest canopy layer and the soil
         is computed as:
         \deqn{g_{a0}= \frac{1}{\frac{U_h}{K_h}\ln(U_{mid}/U_{0})}}{g_a0= 1/((Uh/Kh)*log(U_mid/U_0))}
         where \eqn{U_{mid}}{U_mid} is the wind speed at median cumulated LAI between the top and the soil, and
         \eqn{U_0} the wind speed at soil surface.
}
\examples{
# G_a0 for a coffee plantation managed in agroforestry system:
G_soilcan(Wind=1,ZHT=25,Z_top=24,LAI_top = 0.5,LAI_bot = 4,extwind=0.58)

}
\references{
Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model
            with a multilayer canopy representation for remote sensing purposes. Water
            Resources Research, 1989. 25(5): p. 949-971.
}
\seealso{
\code{\link{G_bulk}} and \code{\link{GetWind}}, which is used internaly.
}
