% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{GBCANMS}
\alias{GBCANMS}
\title{Canopy and Soil boundary layer conductance}
\usage{
GBCANMS(WIND, ZHT, TREEH, Z0 = TREEH * 0.1, ZPD = TREEH * 0.75,
  GBCANMS1MIN = 0.0123, VONKARMAN = Constants()$vonkarman)
}
\arguments{
\item{WIND}{Average daily windspeed (m s-1)}

\item{ZHT}{Wind measurement height (m)}

\item{TREEH}{Average tree height (m)}

\item{Z0}{Roughness length (m), default to 0.1*TREEH}

\item{ZPD}{Zero-plane displacement (m), defaults to 0.75*TREEH}

\item{GBCANMS1MIN}{Minimum allowed atmosphere to canopy conductance (mol m-2 s-1), default to 0.0123}

\item{VONKARMAN}{Con Karman constant, default to \code{Constants()$vonkarman}, 0.41.}
}
\value{
\item{\eqn{GBH}}{Canopy and Soil boundary layer conductance (mol m-2 s-1)}
}
\description{
This function assumes two aerodynamic conductances in series:
 1) from the atmosphere to the canopy, based on Van de Griend (1989), this is actually two conductances,
  one in the inertial sublayer and one in the roughness sublayer.
 2) within the canopy to the soil, based on Choudhury & Monteith (1988).
}
\details{
The defaults for Z0 and ZPD are computed very simply. Other simple formulations:
         1) Lettau (1969) proposed an other way: \deqn{Z0= 0.5 . h* . s / S} where \code{h*}
         is canopy height, s is the average silhouette area (projected area of the tree on a vertical plane)
          and S the specific area, with \deqn{S= A/n}, where A is the total plot area and n the number of trees.
         2) For ZPD, Verhoef et al. (1997) said \deqn{d= 0.67}. h is a good proxy without any prior knowledge.
}
\examples{
# Canop conductance of a forest:
GBCANMS(WIND = 3, ZHT = 26, TREEH = 25)$Canopy
# And the conductance above its soil layer:
GBCANMS(WIND = 3, ZHT = 26, TREEH = 25)$Soil

}
\references{
Van de Griend (1989), Choudhury & Monteith (1988)
}
