% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{GDD}
\alias{GDD}
\title{Daily growing degree days (GDD)}
\usage{
GDD(Tmax = NULL, Tmin = NULL, MinTT = 5, Round = T, Tmean = NULL)
}
\arguments{
\item{Tmax}{Maximum daily temperature (Celsius degree)}

\item{Tmin}{Minimum daily temperature (Celsius degree)}

\item{MinTT}{Minimum temperature threshold, also called base temperature (Celsius degree), default to 5.}

\item{Round}{Boolean. /!\ Important: round the result to 2 decimal, with default to \code{TRUE}.}

\item{Tmean}{Optional. Average daily temperature (Celsius degree). Only needed if Tmax and Tmin are missing.}
}
\value{
\item{GDD}{Growing degree days (Celsius degree)}
}
\description{
Compute the physiological degree days at daily time-step
using the maximum and minimum daily temperature (and optionally using the
average daily temperature, not recommended).
}
\details{
Please keep in mind that this function gives an approximation of the degree days. GDD are
usually computed as the integral of hourly (or less) values.
The round argument is provided for convenience, as growing temperatures with less than 3 digits are likely to be
within the measurement error, and will probably have no visible effect on plant phenology.
Caution, use Tmean only if Tmax and Tmin are not available because it tends to give less powerful estimation.
}
\examples{
# Growing degree days over 10 days :
# Set the seed :
set.seed(1)
GDD(Tmax = rnorm(n = 10, mean = 30, sd = 1),
    Tmin = rnorm(n = 10, mean = 10, sd = 1),
    MinTT = 10)

}
