% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Import.R
\name{Import_Parameters}
\alias{Import_Parameters}
\title{Import model parameters}
\usage{
Import_Parameters(
  path = NULL,
  Names = list(Site = "Site.R", Soil = "Soil.R", Coffee = "Coffee.R", Tree = NULL)
)
}
\arguments{
\item{path}{The path to the parameter files folder. If \code{NULL}, take the default files from the package}

\item{Names}{List of the file names. Default: \code{list(Site="Site.R",Soil="Soil.R",Coffee="Coffee.R",Tree=NULL)}}
}
\value{
A list of all input parameters for DynACof
}
\description{
Import the parameters from local files, or from default (from package data)
}
\details{
For the full list of parameters and the format of the parameter files, see \code{\link[=site]{site()}}.
The function return the parameter files path in the list, to access it, see example.
}
\examples{
# Importing defaults:
Parameters= Import_Parameters()
# Parameters files path:
Parameters$files

}
\seealso{
\code{\link[=DynACof]{DynACof()}} \code{\link[=site]{site()}}
}
