% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Initialization.R
\name{Init_Sim}
\alias{Init_Sim}
\alias{No_Shade.init}
\alias{Tree.init}
\alias{No_Shade.init}
\alias{Tree.init}
\title{Initialise model variables.}
\usage{
Init_Sim(S)

No_Shade.init(S)

Tree.init(S)
}
\arguments{
\item{S}{Model internal working list (= to output list at start)}
}
\description{
Initialise all model variables before start
}
\details{
User should not use this function
}
\keyword{internal}
