% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Metamodels.R
\name{Light_extinction_K}
\alias{Light_extinction_K}
\alias{Metamodels}
\alias{Metamodels}
\alias{Metamodels_soil}
\title{MAESPA metamodels}
\usage{
Light_extinction_K(S, i)

Metamodels(S, i)

Metamodels_soil(S, i)
}
\arguments{
\item{S}{The global list used within the DynACof model.}

\item{i}{The day of interest.}
}
\value{
\item{\eqn{K_{dif_{Tree}}}}{Shade tree diffuse light coefficient}
        \item{\eqn{K_{dif_{Tree}}}}{Shade tree direct light coefficient}
        \item{\eqn{lue_{Tree}}}{Light use efficiency (gC MJ-1)}
        \item{\eqn{T_{Tree}}}{Transpiration (mm d-1)}
        \item{\eqn{H_{Tree}}}{Sensible heat (MJ m-2 d-1)}
}
\description{
Several variables can be highly impacted by the canopy structure. Hence, some variables are
             computed using metamodels from the MAESPA model (see details). These functions are helper
             functions that are made available to the user for the possibility to change the equations easily,
             but not to call them directly.
}
\details{
The purpose of using metamodels in DynACof is to make this two-layer dynamic crop model able to consider
          the 3D canopy heterogeneity effect on fundamental processes, as if it was with a MAESPA-growth model
          coupling. Firstly, the main process impacted by canopy complexity is the light absorbed by the plants
          (Charbonnier et al., 2013). Indeed, in dynamic crop models, the absorbed photosynthetically active
          radiation (APAR) by the canopy is often computed using the simple Beer-Lambert’s law or a derivative,
          with a variable leaf area index in time, but a constant extinction coefficient (Van Oijen et al., 2010b).
          However, heterogeneous canopies such as the shade trees in AFS coffee plantations tends to violate the
          assumption of a constant value for the diffuse (\eqn{K_dif_Tree} the direct (\eqn{K_dir_Tree} light
          extinction coefficient because the spatial distribution of the leaf area is not uniform (high gap fraction)
          and because the leaf area density can change with time through foliage aggregation (Sampson and Smith,
           1993;Sinoquet et al., 2007). Secondly, a comparison between coffee planted in monoculture and under
           agroforestry system showed that canopy complexity affected canopy temperature, water, and energy
           partitioning (Vezy et al., 2018), and probably photosynthesis because it is related to light interception
           and transpiration through stomatal conductance. Therefore, we derived metamodels from MAESPA for the
           diffuse (\eqn{K_dif_Tree} the direct (\eqn{K_dir_Tree} shade tree light extinction coefficients, the light
           use efficiency (LUE, gC MJ-1), the coffee canopy temperature (Tcan, deg Celsius) and leaf water
           potential (MPa), the transpiration (T_\*, mm) and plant sensible heat flux (H_\*). The coffee layer was
           considered homogeneous enough to compute constant extinction coefficients derived from the MAESPA simulation,
           and the partitioning parameter between soil sensible and latent flux was also adjusted using MAESPA outputs.
           MAESPA is a 3D explicit model for energy, carbon and water fluxes simulation, for further details, see:
           \href{https://goo.gl/Z4ehi8}{Vezy et al. (2018)}, or the \href{https://maespa.github.io/}{MAESPA website}
}
\examples{
# Creating a dummy list for use:
S= list(Sim= data.frame(K_Dif_Tree= rep(NA_real_,10),
        K_Dir_Tree= rep(NA_real_,10),LAD_Tree= rnorm(10,3,0.5)))
# Calling the function:
Light_extinction_K(S,1:10)

}
\references{
See \href{https://goo.gl/NVxcVp}{Vezy (2017)}
}
\seealso{
\code{\link{DynACof}}
}
