% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-energy_model.R
\name{light_model_tree}
\alias{light_model_tree}
\alias{light_model_coffee}
\title{Light interception models}
\usage{
light_model_tree(S, i)

light_model_coffee(S, i)
}
\arguments{
\item{S}{The simulation list}

\item{i}{The day index.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link[=DynACof]{DynACof()}} for more details.
}
\description{
Computes the light interception (and transmission) for the shade tree or the coffee.
}
\seealso{
\code{\link[=energy_water_models]{energy_water_models()}}
}
