% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{logistic}
\alias{logistic}
\alias{logistic_deriv}
\title{Logistic function helpers}
\usage{
logistic(xi, u_log, s_log)

logistic_deriv(xi, u_log, s_log)
}
\arguments{
\item{xi}{X vector}

\item{u_log}{Inflexion point (x-value of the sigmoid's midpoint)}

\item{s_log}{Steepness of the curve}
}
\value{
\item{logistic}{Logistic function}
\item{logistic_deriv}{Derivative of the Logistic function}
}
\description{
Compute a logistic function or its derivative
}
\examples{
logistic(1:10,5,0.1)
logistic_deriv(1:10,5,0.1)

}
\seealso{
More informations can be found in \href{https://en.wikipedia.org/wiki/Logistic_function}{
the wikipedia page}
}
