% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-main.R
\name{mainfun}
\alias{mainfun}
\title{Main function}
\usage{
mainfun(cy, Direction, Meteo, Parameters)
}
\arguments{
\item{cy}{The growing cycle (crop rotation)}

\item{Direction}{Simulation directives as a data.frame}

\item{Meteo}{Meteorology data.frame}

\item{Parameters}{Simulation parameters}
}
\value{
The simulation output as a data.frame.
}
\description{
This is the main function of the model that calls all other functions
(meteorology, shade tree, soil) and computes the Coffea simulation. This function
is called by \code{\link[=DynACof]{DynACof()}} under the hood, and users should always call
\code{\link[=DynACof]{DynACof()}} instead of this function because it imports the files,
format the simulation inputs, checks for errors, takes care automatically
of the computation distribution along nodes, and format the outputs.
}
\details{
The Direction \code{data.frame} has to contain at least one column
named Cycle that denotes the crop rotation, Plot_Age, an integer for the plot age,
Plot_Age_num for the plot age as a continuous variable (see example).
}
