% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-Meteo.R
\name{Meteorology}
\alias{Meteorology}
\title{Import Meteorology for model input}
\usage{
Meteorology(file = NULL, Period = NULL, Parameters = Import_Parameters())
}
\arguments{
\item{file}{Either the file name to read, a shell command that preprocesses the file (e.g. fread("grep blah filename"))
or the input itself as a string, see \code{\link[data.table]{fread}}. In both cases, a length 1 character string.
A filename input is passed through path.expand for convenience and may be a URL starting http:// or file://.
Default to \code{NULL}, which make the function return the \code{Aquiares} package example data.}

\item{Period}{The desired time period to be returned in the form of a vector of two POSIX dates (min and max).}

\item{Parameters}{A list of parameters:
\itemize{
  \item Start_Date: optional, the date of the first meteo file record. Only needed if the Date column is missing (Posixct).
  \item FPAR      : Fraction of global radiation that is PAR (source: MAESPA model), only needed if either RAD or PAR is missing (fraction).
  \item Elevation : elevation of the site, only needed if atmospheric pressure is missing (m)
  \item Latitude  : latitude of the site, only needed if the diffuse fraction of light is missing (degree)
  \item WindSpeed : constant wind speed, only needed if windspeed is missing (m s-1)
  \item CO2       : constant atmospheric \eqn{CO_2} concentration, only needed if \eqn{CO_2} is missing (ppm)
  \item MinTT     : minimum temperature threshold for degree days computing, see \code{\link{GDD}}
  \item MaxTT     : maximum temperature threshold for degree days computing, see \code{\link{GDD}}
  \item albedo    : site shortwave surface albedo, only needed if net radiation is missing, see \code{\link{Rad_net}}
}}
}
\value{
A daily timestep meteorology data.frame (invisibly) with either the data read above or computed.
}
\description{
This function aims to import the meteorology data, check it, and compute missing variables.
}
\details{
The imported file is expected to be at daily time-step.
         The albedo is used to compute the system net radiation that is then used to compute the soil net radiation using an
         extinction coefficient with the plot LAI following the Shuttleworth & Wallace (1985) formulation. This computation is
         likely to change in the near future to add a more uniform process-based formulation (such as Choudhury & Monteith 1988).
\tabular{llll}{\strong{Var} \tab \strong{unit} \tab \strong{Definition} \tab \strong{If missing} \cr
Date            \tab POSIXct date\tab Date in POSICct format                       \tab Computed from start date parameter, or set a dummy date if missing\cr
year            \tab year        \tab Year of the simulation                       \tab Computed from Date \cr
DOY             \tab day         \tab day of the year                              \tab Computed from Date \cr
Rain            \tab mm          \tab Rainfall                                     \tab Assume no rain \cr
Tair            \tab deg C       \tab Air temperature (above canopy)               \tab Computed from Tmax and Tmin \cr
Tmax            \tab deg C       \tab Maximum air temperature durnig the day       \tab Required (error) \cr
Tmin            \tab deg C       \tab Minimum air temperature durnig the day       \tab Required (error) \cr
RH              \tab \%          \tab Relative humidity                            \tab Not used, but prefered over VPD for Rn computation \cr
RAD             \tab MJ m-2 d-1  \tab Incident shortwave radiation                 \tab Computed from PAR \cr
Pressure        \tab hPa         \tab Atmospheric pressure                         \tab Try to compute from VPD, Tair and Elevation, or Tair and Elevation. \cr
WindSpeed       \tab m s-1       \tab Wind speed                                   \tab Try to set it to constant: \code{Parameters$WindSpeed} \cr
CO2             \tab ppm         \tab Atmospheric CO2 concentration                \tab Try to set it to constant: \code{Parameters$CO2}\cr
DegreeDays      \tab deg C       \tab Growing degrre days                          \tab Computed using \code{\link{GDD}} \cr
PAR             \tab MJ m-2 d-1  \tab Incident photosynthetically active radiation \tab Computed from RAD \cr
FDiff           \tab Fraction    \tab Diffuse light fraction                       \tab Computed using \code{\link{Diffuse_d}} usinf Spitters formula \cr
VPD             \tab hPa         \tab Vapor pressure deficit                       \tab Computed from RH \cr
Rn              \tab MJ m-2 d-1  \tab Net radiation (will soon be depreciated)     \tab Computed using \code{\link{Rad_net}} with RH, or VPD \cr
DaysWithoutRain \tab day         \tab Number of consecutive days with no rainfall  \tab Computed from Rain \cr
Air_Density     \tab kg m-3      \tab Air density of moist air (\eqn{\rho}) above canopy \tab Computed using \code{\link[bigleaf]{air.density}}}
         It is highly recommended to set the system environment timezone to the one from the meteorology file.
         For example the default meteorology file (\code{\link{Aquiares}}) has to be set to \code{Sys.setenv(TZ="UTC")}.
}
\examples{
\dontrun{
if(interactive()){
 Sys.setenv(TZ="UTC")
 Met_c= Meteorology()

 # Get the units of the output variables:
 attr(Met_c,"unit")
 }
}

}
\seealso{
\code{\link{DynACof}}
}
\author{
R. Vezy; O. Roupsard
}
