% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{Paliv_dis}
\alias{Paliv_dis}
\title{Continuous percentage of living tissue distribution}
\usage{
Paliv_dis(Age_Max, P_Start, P_End, k)
}
\arguments{
\item{Age_Max}{Maximum age of the organ (year)}

\item{P_Start}{Percentage of living tissue at first age (\% of dry mass)}

\item{P_End}{Percentage of living tissue at last age (\% of dry mass)}

\item{k}{Rate between P_Start and P_End}
}
\value{
Living tissue at each age in \% of organ dry mass
}
\description{
Computes the percentage of living tissue in the organ according to age
}
\details{
The percentage of living tissue is computed as follows:
 \deqn{P_{End}+\left((P_{Start}-P_{End})\cdot e^{seq(0,-k,length.out=Age_{Max})}\right)}{
       P_End+((P_Start-P_End)*exp(seq(0,-k,length.out = Age_Max)))}
}
\examples{
Paliv_dis(40,1,0.05,5)

}
