% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{plot.stacked}
\alias{plot.stacked}
\title{x' - a vector of values
y' - a matrix of data series (columns) corresponding to x
order.method' = c("as.is", "max", "first")
col' - fill colors for polygons corresponding to y columns (will recycle)
border' - border colors for polygons corresponding to y columns (will recycle) (see ?polygon for details)
lwd' - border line width for polygons corresponding to y columns (will recycle)
...' - other plot arguments}
\usage{
\method{plot}{stacked}(x, y, order.method = "as.is", ylab = "", xlab = "",
  border = NULL, lwd = 1, col = rainbow(length(y[1, ])), ylim = NULL,
  xlim = NULL, density = -1, lty = 1, Fignum = NULL, ...)
}
\description{
x' - a vector of values
y' - a matrix of data series (columns) corresponding to x
order.method' = c("as.is", "max", "first")
col' - fill colors for polygons corresponding to y columns (will recycle)
border' - border colors for polygons corresponding to y columns (will recycle) (see ?polygon for details)
lwd' - border line width for polygons corresponding to y columns (will recycle)
...' - other plot arguments
}
