% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{previous_i}
\alias{previous_i}
\title{Find the ith previous index}
\usage{
previous_i(x, n_prev)
}
\arguments{
\item{x}{Current index}

\item{n_prev}{Target number of indexes before x}
}
\description{
Find the ith previous index while avoiding 0 or negative indexes.
}
\details{
This function is used to find the nth previous index without making an error with negative
or 0 index.
}
\examples{
# Find the 10th index before 15:
previous_i(15,10)
# Find the 10th index before 5:
previous_i(5,10)
# should return 1 (no index before 1)

}
\keyword{internal}
