% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{Rad_ext}
\alias{Rad_ext}
\title{Daily extra-terrestrial radiation}
\usage{
Rad_ext(DOY, Latitude, Gsc = Constants()$Gsc)
}
\arguments{
\item{DOY}{Ordinal date (integer): day of year from 1st January (day)}

\item{Latitude}{Latitude (deg)}

\item{Gsc}{The solar constant (W m-2), default to \code{Constants()$Gsc} (= 1367).}
}
\value{
\eqn{S0}, the daily extra-terrestrial radiation (\eqn{MJ\ m^{-2}\ d^{-1}}{MJ m-2 d-1})
}
\description{
Compute the daily extra-terrestrial radiation at a plane parallel to the
             earth surface (\eqn{S0} or \eqn{H0} depending on the source)
             following Khorasanizadeh and Mohammadi (2016).
}
\examples{
# Daily extra-terrestrial radiation on january 1st at latitude 35 N :
Rad_ext(DOY= 1,Latitude= 35)

}
\references{
Khorasanizadeh, H. and K. Mohammadi, Diffuse solar radiation on a horizontal surface:
            Reviewing and categorizing the empirical models. Renewable and Sustainable Energy Reviews,
            2016. 53: p. 338-362.
}
