% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{Rad_net}
\alias{Rad_net}
\title{Daily net radiation}
\usage{
Rad_net(DOY, RAD, Tmax, Tmin, VPD, Rh = NULL, Latitude, Elevation, albedo,
  type_ea = c("VPD", "Temp"), sigma = Constants()$sigma,
  Gsc = Constants()$Gsc)
}
\arguments{
\item{DOY}{Day Of Year from 1st January (day)}

\item{RAD}{Incident daily total radiation (MJ m-2 d-1)}

\item{Tmax}{Maximum daily air temperature (celsius degree)}

\item{Tmin}{Minimum daily air temperature (celsius degree)}

\item{VPD}{Mean daily Vapor Pressure Deficit (hPa), only needed if \code{Rh} is missing}

\item{Rh}{Average daily relative humidity (\code{\%})}

\item{Latitude}{Latitude (deg)}

\item{Elevation}{Elevation (m)}

\item{albedo}{Shortwave surface albedo (-)}

\item{sigma}{Stefan-Boltzmann constant (W m-2 K-4), default to \code{Constants()$sigma}.}

\item{Gsc}{Solar constant (W m-2), default to \code{Constants()$Gsc} (=1367).}

\item{RAD}{Incident total radiation (MJ m-2 d-1)}
}
\value{
\item{\eqn{Rn}}{Daily net radiation (MJ m-2 d-1)}
}
\description{
Compute the daily net radiation of the system using incident radiation, air
             temperature, wind speed, relative humidity and the albedo. A clear description
             of this methodology can be found in Allen et al. (1998) or in An et al. (2017).
}
\details{
The daily net radiation is computed using the surface albedo. This method is only a
         simple estimation. Several parameters (ac, bc, a1 and b1) are taken from
         Evett et al. (2011). The net radiation is computed as:
         \deqn{Rn= (1-albedo)*RAD-(ac*(RAD/Rso)+bc)*(a1+b1*ea^0.5)*sigma*((Tmax^4+Tmin^4)/2)}
         And is derived from the equation :
         \deqn{Rn= (1-albedo)*RAD-Rln}
         where \eqn{Rln} is the net upward longwave radiation flux.
         The actual vapor pressure \deqn{ea} can be computed using either VPD or the relative
         humidity and the maximum and minimum daily temperature. If both are provided, Rh will
         be used.
}
\examples{
# Daily net radiation on january 1st at latitude 9 N :
Rad_net(DOY= 1,RAD= 5,Tmax= 16,Tmin= 10,VPD=1.05,
        Latitude=9,Elevation=1000,albedo=0.146)

}
\references{
An, N., S. Hemmati, and Y.-J. Cui, Assessment of the methods for determining net
            radiation at different time-scales of meteorological variables. Journal of Rock
            Mechanics and Geotechnical Engineering, 2017. 9(2): p. 239-246.
}
