% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Tree_DCM.R
\name{Shade.Tree}
\alias{Shade.Tree}
\alias{No_Shade}
\alias{No_Shade}
\title{Shade Tree module}
\usage{
Shade.Tree(S, i)

No_Shade(...)
}
\arguments{
\item{S}{The main simulation list to make the computation on and to modify.}

\item{i}{The index of the day since the first day of the simulation.}
}
\value{
Nothing, modify the list of simulation \code{S} in place. See \code{\link{DynACof}} for
        more details.
}
\description{
This make all computations for shade trees (similar to coffee, but no fruits) for the ith
             day by modifying the \code{S} list in place.
}
\note{
This function shouldn't be called by the user. It is made as a module so it is easier for
      advanced users to modify the code.
      \code{No_Shade()} is used as an empty function that is called when there are no shade trees.
}
\seealso{
\code{\link{DynACof}}
}
\keyword{internal}
