% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\docType{class}
\name{Simulation-class}
\alias{Simulation-class}
\alias{SimulationClass}
\title{Generator for class Simulation}
\value{
The generator for the simulation class.
}
\description{
Set the reference class for the simulation object, which hold the Meteorology, the simulation outputs and the
             parameters.
}
\details{
This function is internal and shouldn't be called by users.
}

\examples{
# Create an object of class "Simulation":
S= SimulationClass$new()
# Fill in the parameters:
S$Parameters= Parameters
}
\keyword{internal}
