% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Soil_DCM.R
\name{Soilfun}
\alias{Soilfun}
\title{Soil module subroutine}
\usage{
Soilfun(S, i)
}
\arguments{
\item{S}{The main simulation list to make the computation on and to modify.}

\item{i}{The index of the day since the first day of the simulation.}
}
\value{
Modify the list of simulation \code{S} in place. See \code{\link{DynACof}} for
        more details.
}
\description{
Make all computations for soil water balance for the ith
             day by modifying the \code{S} list in place.
}
\note{
This function shouldn't be called by the user. It is made as a subroutine so it is easier for
      advanced users to modify the code.
}
\seealso{
\code{\link{DynACof}}
}
\keyword{internal}
