% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{Sucrose_cont_perc}
\alias{Sucrose_cont_perc}
\title{Fruit sucrose accumulation}
\usage{
Sucrose_cont_perc(x, a, b, x0, y0)
}
\arguments{
\item{x}{Cumulated degree days}

\item{a}{Parameters}

\item{b}{Parameters}

\item{x0}{Mid-maturation (logistic function inflexion point)}

\item{y0}{Sucrose content at the beginning (in \%, 1-100)}
}
\value{
\item{Sucrose content}{(In \% of fruit total dry mass)}
}
\description{
Logistic sucrose accumulation into coffee fruits through time
}
\examples{
Sucrose_cont_perc(1:10,5.3207,-28.5561,191,3.5)

}
\references{
Pezzopane et al. (2012) :
Pezzopane, J., et al., Agrometeorological parameters for prediction of the maturation period
of Arabica coffee cultivars. International Journal of Biometeorology, 2012. 56(5): p. 843-851.
}
