% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-test_parameters.R
\name{test_parameters}
\alias{test_parameters}
\title{Tests on input parameters}
\usage{
test_parameters(Parameters, isTree = FALSE)
}
\arguments{
\item{Parameters}{The full list of parameter values, generally output from \code{\link[=Import_Parameters]{Import_Parameters()}}}

\item{isTree}{Boolean. Set to \code{TRUE} if shade trees are simulated.}
}
\value{
An error if any issue is encountered
}
\description{
Different tests for the user-inputed parameter files:
\describe{
\item{Mandatory parameter}{Test if all mandatory parameters are present in the files}
\item{Parameter values}{Some parameter values are checked before running}
}
}
\examples{
# No error:
test_parameters(Parameters= Import_Parameters())

\donttest{
 # Removing a mandatory parameter from the soil:
 params= Import_Parameters()
 params$REWc= NULL
 test_parameters(Parameters= params)
}
}
