% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varnames}
\alias{varnames}
\title{Output variables}
\format{A data frame with three columns:
\describe{
  \item{Type}{Variable name}
  \item{Unit}{unit of the variable}
  \item{Definition}{a brief definition of the variable}
}
For the moment the model has 106 output variables.}
\usage{
varnames
}
\description{
Generic output variables from DynACof, with their units and brief description
}
\keyword{datasets}
