% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{warn.var}
\alias{warn.var}
\title{Warn or stop execution if mandatory meteorology input variables are not provided}
\usage{
warn.var(Var, replacement, type = "error")
}
\arguments{
\item{Var}{Input variable name}

\item{replacement}{Replacement variable that is used to compute \code{"Var"}}

\item{type}{Type of error to return : either}
}
\description{
Help to explain which variable is missing and/or if there are replacements
}
\note{
This function helps to debug the model when some mandatory meteorological variables
are missing from input: either an error (default), or a warning.
If \code{"replacement"} is not provided in the meteorology file either, this function
will return an error with a hint on which variables can be provided to compute
\code{"Var"}
}
\keyword{internal}
