% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{write.results}
\alias{write.results}
\title{Write output results to disk}
\usage{
write.results(
  FinalList,
  output = ".RData",
  Simulation_Name = NULL,
  Outpath = "Outputs",
  ...
)
}
\arguments{
\item{FinalList}{The model output list}

\item{output}{Output format. Character. \code{".RData"} if single file, or anything else for
3 separate file. Default: \code{".RData"}}

\item{Simulation_Name}{The name of the simulation. Used for the name of the outputs. Default: \code{NULL}.}

\item{Outpath}{The path to the folder to write on.}

\item{...}{Further parameters to pass to \code{\link[data.table:fwrite]{data.table::fwrite()}}. Only used if
\code{output!=".RData"}.}
}
\description{
Write the ouptut list to disk either under a unique \code{.RData} file or three separated
files:
\itemize{
\item Model Output (csv file)
\item Meteoroloy output (input to the model, csv file)
\item Model parameters (txt file)
}
}
\details{
To re-import the output, the user should use \code{\link[base:readRDS]{base::readRDS()}}
}
