% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{CB}
\alias{CB}
\title{Temperature-dependent correction (CB)}
\usage{
CB()
}
\value{
A function to predict the temperature correction coefficient (see Eq. 36 from Vezy et al. (in prep.))
}
\description{
Make a monotone Hermite spline function (Fritsch and Carlson, 1980) fitted on Drinnan et al. (1995) data.
}
\details{
As temperature increases, the number of nodes on coffee increases due to increased vegetative
growth, but the number of buds per nodes decreases. This is computed by using a temperature correction
factor that decrease with increasing mean temperature during bud development (0-1, and =1 if mean T < 23).
This factor is then applied on the number of buds that break dormancy (less buds break dormancy with
increasing T).
}
\examples{
# The function returns the temperature correction function itself.
# Exemple of corrcetion factor for a 32 Celsius degrees temperature:
CB()(32)

}
\references{
\itemize{
  \item Fritsch, F. N. and R. E. Carlson (1980). "Monotone Piecewise Cubic Interpolation."  17(2): 238-246.
  \item Drinnan, J. and C. Menzel, Temperature affects vegetative growth and flowering of coffee (Coffea arabica L.).
  Journal of Horticultural Science, 1995. 70(1): p. 25-34.
}
}
