% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-main.R
\name{dynacof_i}
\alias{dynacof_i}
\title{Step-by-step DynACof}
\usage{
dynacof_i(i, S, verbose = TRUE)
}
\arguments{
\item{i}{Either an integer, or a range giving the day of simulation needed. Match the row index, so \code{i=1} make
a simulation for the first row of Sim and Met.}

\item{S}{The simulation list (see \code{\link[=DynACof]{DynACof()}}).}

\item{verbose}{Boolean. Prints progress bar if \code{TRUE} (default).}
}
\value{
The modified simulation list \code{S}
}
\description{
Using DynACof one iteration after another. Allows to run a DynACof simulation with starting
at age > 0 with initializations.
}
\examples{
\dontrun{
Making a regular simulation using example data:
S= DynACof(Period= as.POSIXct(c("1979-01-01", "1980-12-31")))

# Value of the maintenance respiration for coffee on day i=100:
i= 100
S$Sim$Rm[i]

# Changing the value of Tair in the meteorology for day 100:
S$Meteo$Tair[i]= S$Meteo$Tair[i]+10.0

S= dynacof_i(i,S)

# New value of the maintenance respiration for coffee:
S$Sim$Rm[i]

# To re-run DynACof for several days, use a range for `i`:
S= dynacof_i(i:(i+10),S)

}
}
