% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{G_interlay}
\alias{G_interlay}
\title{Canopy layer to canopy layer aerodynamic conductance}
\usage{
G_interlay(
  Wind,
  ZHT,
  Z_top,
  Z0 = Z_top * 0.1,
  ZPD = Z_top * 0.75,
  alpha = 1.5,
  ZW = ZPD + alpha * (Z_top - ZPD),
  LAI_top,
  LAI_bot,
  extwind = 0,
  vonkarman = Constants()$vonkarman
)
}
\arguments{
\item{Wind}{Average daily wind speed above canopy (m s-1)}

\item{ZHT}{Wind measurement height (m)}

\item{Z_top}{Average canopy height of the taller crop (m)}

\item{Z0}{Roughness length (m). Default: \code{0.1*Z_top}}

\item{ZPD}{Zero-plane displacement (m), Default: \code{0.75*Z_top}}

\item{alpha}{Constant for diffusivity at top canopy. Default: \code{1.5} following
Van de Griend et al (1989).}

\item{ZW}{Top height of the roughness sublayer (m). Default: \code{ZPD+alpha*(Z_top-ZPD)}}

\item{LAI_top}{Leaf area index of the upper layer (m2 leaf m-2 soil).}

\item{LAI_bot}{Leaf area index of the layer below the upper layer (m2 leaf m-2 soil).}

\item{extwind}{Extinction coefficient. Default: \code{0}, no extinction.}

\item{vonkarman}{Von Karman constant, default to \code{Constants()$vonkarman}, 0.41.}
}
\value{
\item{g_af}{The aerodynamic conductance of the air between two canopy layers (m s-1)}
}
\description{
Compute the aerodynamic conductance for sensible and
latent heat between canopy layers following Van de Griend
and Van Boxel (1989).
}
\details{
\code{alpha} can also be computed as:
\deqn{alpha=\frac{zw-d}{Z_{top}-d}}{alpha= (zw-d)/(Z_top-d)}
The aerodynamic conductance between canopy layers is computed as:
\deqn{g_{af}= \frac{1}{\frac{U_h}{K_h}\ln(U_{mid}/U_{inter})}}{g_af= 1/((Uh/Kh)*log(U_mid/U_inter))}
where usually \eqn{U_{mid}}{U_mid} is the wind speed at (median) cumulated LAI between
the top and the soil, and \eqn{U_{inter}}{U_inter} the wind speed at the height between
the two canopy layers. In this function, \eqn{U_{mid}}{U_mid} and \eqn{U_{inter}}{U_inter} are computed
relative to the leaf area instead of the height of the vegetation layers.
}
\examples{
# G_af for a coffee plantation managed in agroforestry system:
G_interlay(Wind = 3,ZHT = 25,Z_top = 2,LAI_top = 0.5,LAI_bot = 4)


}
\references{
Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model
with a multilayer canopy representation for remote sensing purposes. Water
Resources Research, 1989. 25(5): p. 949-971.
}
\seealso{
\code{\link[=G_bulk]{G_bulk()}} and \code{\link[=GetWind]{GetWind()}}, which is used internaly.
}
