% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{GetWind}
\alias{GetWind}
\title{Get the average wind speed at center of canopy layer}
\usage{
GetWind(Wind, LAI_lay, LAI_abv, extwind = 0)
}
\arguments{
\item{Wind}{Above canopy wind speed (m s-1)}

\item{LAI_lay}{Leaf area index of the layer (m2 leaves m-2 soil)}

\item{LAI_abv}{Cumulated leaf area index above the layer (m2 leaves m-2 soil)}

\item{extwind}{Extinction coefficient. Default: \code{0}, no extinction.}
}
\value{
\item{WindLay}{The winspeed at the center of the layer (m s-1)}
}
\description{
Calculate the wind speed decrease with increasing canopy depth using an
             exponential extinction coefficient and a cumulated LAI above the target point.
}
\details{
As the function computes the average wind speed at the center of the layer, it uses
         the \code{LAI_lay} parameter to add half of the target layer to the cumulated LAI above:
         \deqn{WindLay=Wind*e^{^{\left(-extwind*\left(LAI_{abv}+\frac{LAI_{lay}}{2}\right)\right)}}}{
         WindLay= Wind*e(-extwind*(LAI_abv+LAI_lay/2)}
}
\examples{
# Windspeed in a coffee layer managed in agroforestry system
GetWind(Wind=3,LAI_lay=4,LAI_abv=0.5,extwind= 0.58)

}
\references{
\href{https://maespa.github.io}{MAESPA model}
}
