% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Functions.R
\name{PENMON}
\alias{PENMON}
\title{Evapotranspiration}
\usage{
PENMON(Rn, Wind, Tair, ZHT, TREEH, Pressure, Gs, VPD,
  Parameters = Constants())
}
\arguments{
\item{Rn}{Net radiation (MJ m-2 d-1)}

\item{Wind}{Wind speed (m s-1)}

\item{Tair}{Air temperature (Celsius degree)}

\item{ZHT}{Wind measurement height (m)}

\item{TREEH}{Average tree height (m)}

\item{Pressure}{Atmospheric pressure (hPa)}

\item{Gs}{Stomatal conductance (mol m-2 s-1)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{Parameters}{Constant parameters, default to Constants(), if different values are needed:
Cp - specific heat of air for constant pressure (J K-1 kg-1) \cr
Rgas - universal gas constant (J mol-1 K-1) \cr
Kelvin - conversion degree Celsius to Kelvin \cr
H2OMW - conversion from kg to mol for H2O (kg mol-1)}
}
\value{
\item{\eqn{ET}}{Daily evapotranspiration (mm d-1)}
}
\description{
Compute the daily evaporation or transpiration of the surface using the
             Penman-Monteith equation.
}
\details{
The daily evapotranspiration is computed using the Penman-Monteith equation, and a
         set of conductances as :
         \deqn{ET= (Delta * Rn*10^6 + rho * Cp * (VPD/10) * GH) / (Delta + gamma * (1 + GH / GV))/\lambda}
         where \eqn{\Delta} is the slope of the saturation vapor pressure curve (kPa K-1),
         \eqn{\rho} is the air density (kg m-3), GH the canopy boundary layer conductance (m s-1),
         \eqn{\gamma} the psychrometric constant (kPa K-1) and \eqn{GV} the boundary + stomatal
         conductance to water vapour (m s-1). To simulate evaporation, \eqn{Gs} can be set
         to nearly infinite (e.g. \eqn{Gs= 1E09}).
}
\examples{
# leaf evaporation of a forest :
PENMON(Rn= 12, Wind= 0.5, Tair= 16, ZHT= 26, TREEH= 25, Pressure= 900, Gs = 1E09, VPD= 2.41)

}
\references{
Allen R.G., Pereira L.S., Raes D., Smith M., 1998: Crop evapotranspiration -
             Guidelines for computing crop water requirements - FAO Irrigation and drainage
             paper 56.
}
\seealso{
\code{\link[bigleaf]{reference.ET}} and \href{https://maespa.github.io/}{MAESPA model}
}
