% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Initialization.R
\docType{class}
\name{Simulation-class}
\alias{Simulation-class}
\alias{SimulationClass}
\title{Generator for class Simulation (simulation object)}
\description{
Generator for class Simulation (simulation object)
}
\section{Fields}{

\describe{
\item{\code{Sim}}{A list, this is the simulation output list}

\item{\code{Met_c}}{A list, this is the input meteorology}

\item{\code{Parameters}}{A list, this is the parameters list}

\item{\code{Zero_then_One}}{A vector, this is a helper vector (will be depreciated soon)}
}}


