% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Site.R, R/0-Soil.R, R/0-Coffee.R, R/0-Tree.R
\name{site}
\alias{site}
\alias{soil}
\alias{coffee}
\alias{Tree}
\title{Default (example) parameter functions}
\usage{
site()

soil()

coffee()

Tree()
}
\value{
A list of parameters needed for a DynACof simulation.
}
\description{
These functions return the default parameter values for DynACof.
They are mainly used invisibly as defaults from \code{\link[=Import_Parameters]{Import_Parameters()}}, but can still be
called by the user for conveniance (but not needed for a model run). The parameters are divided into
four functions: \code{site()}, \code{soil()}, \code{coffee()}, and \code{Tree()}.
}
\details{
These input functions are stored into R-script that can be customized at will by the user to fit
new conditions. In that case, the user have to write the resulting functions to separate files
corresponding to the site, the soil, the coffee layer and the tree layer.
Then \code{\link[=DynACof]{DynACof()}} read them using the \code{Inpath} and \code{FileName} parameters.
File templates with default parameter values and an example use are available from this separate
\href{https://github.com/VEZY/DynACof_inputs}{github repository}
}
\note{
The Tree parameter file is optional. All other parameter files should have the same structure as
the one from the default parameter files. Other parameters can be added to the files though.
}
\section{site}{

The default site is a stand from the Aquiares farm, located in Costa Rica. It is a \emph{Coffea arabica}
plantation in agroforestry management under \emph{Erythrina poeppigiana} shade trees. The plot is visible
at this \href{https://goo.gl/7FRNXg}{address}, and a full desciption is available
\href{https://www.researchgate.net/publication/323398728_Measuring_and_modelling_energy_partitioning_in_canopies_of_varying_complexity_using_MAESPA_model}{here}
and \href{https://www.researchgate.net/publication/323469257_Simulation_de_pratiques_de_gestion_alternatives_pour_l'adaptation_des_plantations_perennes_aux_changements_globaux}{here}.
}

\section{soil}{

The default soil parameters.
}

\section{coffee}{

The default coffee parameters. It is essentially a high density plantation (6666 coffee plants
per hectares) of \emph{Coffea arabica var. Caturra} that are pruned every year to sustain
three resprouts per stump in average.
}

\section{Tree}{

The default shade tree parameters. The shade trees in Aquiares were  \emph{Erythrina poeppigiana}
planted at high density (250 trees ha-1) and pruned to optimize light transmitted to the \emph{Coffea},
and were depressed in 2000 to a low density of ~7.4 trees ha-1.
From 2000, these trees made a relatively large crown with an average height of 26 m in 2018
on this site. NB: the Tree parameter file is optional, and not needed
for monospecific coffee plantations.
}

\seealso{
\code{\link[=DynACof]{DynACof()}}
}
